Add `name` to list of ignored function attributes
Pacomo transformer fails on Microsoft Edge because it tries to copy the `name` property of the function being wrapped. This is because `Object.getOwnPropertyNames` includes `name` on the named function but not on `function () {}`, so `name` is not in the banned list `ignoredFunctionStatics`.

Chrome works because `function () {}` has a `name` property, IE11 works for the opposite reason─the source function lacks a `name` property!

Suggested fix is to add `name` to the additional hard-coded banned list.